requirements.txt
paramiko
gitpython